+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Version 0.24
 - added preliminary `dlreplacecontent()` function to replace the content of a linked file.
 - `dl_directory` check constraint that dirpath is like '/%/'
 - new `is_local()`, `is_valid()` as `is_http_success()` immutable functions, for use in check constraints and such
-- add `read()` function to read binary files
+- add `read()` and `write()` functions to read and write binary files
 - forbid * in file names
 - new `iri()` function to convert IRIs (unicode) to URIs (escaped) 
 - `dlvalue()` now supports `IRI` link type, bug fixes
@@ -12,7 +12,7 @@ Version 0.24
 - added `datalink.http_response_codes` reference table 
 - added tables `datalink.dl_access_web` and `datalink.access_web` table and view 
 - added preliminary `has_web_privilege()` functions
-- simplify file: URLs without host
+- simplify `file:` URLs to be without host if not needed
 
 Version 0.23
 ------------